requirements.txt
pytest
pylint
jupyter
pytest-cov
pandas
nbval
click
flask
requests
matplotlib
google-cloud
google-cloud-bigquery
bq_helper
